# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Daily TCG Wallpaper API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallpaper {
  id: number;
  title: string;
  /** Emotional tone e.g. Ethereal, Fierce, Serene, Mystic */
  mood: string;
  /** Visual style e.g. Holographic, Dark Fantasy, Neon Arcane, Celestial */
  style: string;
  imageUrl: string;
  isPremium: boolean;
  /** ISO date string */
  releaseDate: string;
}

export interface FavoriteResult {
  wallpaperId: number;
  favorited: boolean;
}

export interface ErrorResponse {
  error: string;
}