# lib/api-zod/src/generated/types/favoriteResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Daily TCG Wallpaper API
 * OpenAPI spec version: 0.1.0
 */

export interface FavoriteResult {
  wallpaperId: number;
  favorited: boolean;
}